feat: Added missing features: confirmation dialog, posts pagination in feed and some fixed styles

+++ src/app/features/comments/components/comment-list/comment-list.component.ts
@@ -4,8 +4,11 @@ import { MatCardModule } from '@angular/material/card';
 import { MatListModule } from '@angular/material/list';
 import { MatInputModule } from '@angular/material/input';
 import { MatIconModule } from '@angular/material/icon';
+import { MatDialog } from '@angular/material/dialog';
 import { FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
+import { ConfirmDialogComponent } from '../../../../shared/components/confirm-dialog/confirm-dialog.component';
 import { CommentService, Comment } from '../../../../core/services/comment.service';
+import { MatButtonModule } from '@angular/material/button';
 
 @Component({
   selector: 'app-comment-list',
@@ -13,6 +16,7 @@ import { CommentService, Comment } from '../../../../core/services/comment.servi
     CommonModule,
     MatCardModule,
     MatListModule,
+    MatButtonModule,
     MatIconModule,
     MatInputModule,
     ReactiveFormsModule,
@@ -24,6 +28,7 @@ export class CommentListComponent implements OnInit {
   @Input() postId!: string;
   comments: Comment[] = [];
   private commentService = inject(CommentService);
+  private dialog = inject(MatDialog);
   loggedUserId: string = localStorage.getItem('userId') || '';
   editingCommentId: string | null = null;
   editForm: FormGroup;
@@ -66,8 +71,16 @@ export class CommentListComponent implements OnInit {
   }
 
   deleteComment(commentId: string) {
-    this.commentService.deleteComment(commentId).subscribe(() => {
-      this.comments = this.comments.filter((comment) => comment._id !== commentId);
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: { message: 'Are you sure you want to delete this comment?' },
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.commentService.deleteComment(commentId).subscribe(() => {
+          this.comments = this.comments.filter((comment) => comment._id !== commentId);
+        });
+      }
     });
   }
 }
